backend/requirements.txt
# FastAPI and web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis and Celery
redis==5.2.0
celery==5.4.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.9.0

# Validation and serialization
pydantic[email]==2.10.0
pydantic-settings==2.6.1
email-validator==2.2.0

# AI APIs
anthropic==0.39.0
openai==1.54.4

# External APIs
httpx==0.27.2

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.3

# CORS
python-dotenv==1.0.1
